+++ NEWS.md
@@ -12,3 +12,6 @@
 ### CytoMDS 0.99.3
 - new version of computeMetricMDS() which automatically sets 
 the number of dimensions to reach a target *pseudo R squared*
+- added ggplotly() functionality for output MDS plots
+- in `ggplotSampleMDS()`, added `flipXAxis`, `flipYAxis` 
+to possibly ease low dimensional projection comparisons